+++ crates/api/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Patch
 
+- Make sure at compile-time that exactly one `host` or `wasm` feature is enabled
 - Update dependencies
 
 ## 0.6.1