+++ crates/protocol-usb/CHANGELOG.md
@@ -21,4 +21,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 4 -->
+<!-- Increment to skip CHANGELOG.md test: 5 -->